src/shared/auth/pyproject.toml
[tool.poetry]
name = "wine-fermentation-auth"
version = "0.1.0"
description = "Shared authentication module for wine fermentation system"
authors = ["Wine Fermentation Team"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.0"}
asyncpg = "^0.29.0"
pyjwt = "^2.10.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
fastapi = "^0.120.4"
python-multipart = "^0.0.20"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
pytest-cov = "^4.1.0"
aiosqlite = "^0.21.0"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "strict"
addopts = "-v --tb=short"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
